refactor(type): Trim the include graph of the simple-function headers
Defining a Velox simple function currently costs every translation unit
the entire runtime type system. `functions/Macros.h` -- included by every
`*Functions.h` header -- pulled in `type/Type.h` twice over: once
directly, and once transitively through `type/SimpleFunctionApi.h` ->
`type/CppToType.h`. `Type.h` is ~2400 lines and brings folly/Random.h,
folly/container/F14Set.h, common/serialization/Serializable.h, glog,
gflags, <iomanip> and the virtual `Type` hierarchy with it.

None of that is needed to define a simple function. Three changes:

1. `Macros.h`: drop the direct `#include "velox/type/Type.h"`. Every tag
   the `VELOX_DEFINE_FUNCTION_TYPES` macro names (`Map`, `Array`, `Row`,
   `Varchar`, `Varbinary`) is defined in `SimpleFunctionApi.h`, not in
   `Type.h`.

2. Split `SimpleFunctionApi.h` (499 lines). Its only include,
   `CppToType.h`, was needed exclusively by the `CppToType` /
   `SimpleTypeTrait` / `MaterializeType` specializations that begin at
   line 296; the signature tag types above them need nothing from it.
   The tags move to a new `type/SimpleFunctionTags.h` whose heaviest
   dependency is `<fmt/format.h>`, and `SimpleFunctionApi.h` becomes an
   umbrella that includes the tags plus `CppToType.h` and keeps the
   specializations. Existing includers of `SimpleFunctionApi.h` are
   unaffected. `Macros.h` now includes only the tags header.

   One trait has to move for the tags header to stand alone:
   `isOpaqueType` derives from `is_shared_ptr`, which #18365 consolidated
   into `Type.h`. Reaching it from `SimpleFunctionTags.h` would undo the
   whole point of the split, so the four-line trait moves to `TypeKind.h`
   alongside the rest of the compile-time type metadata. It depends only
   on <memory>, and `Type.h` includes `TypeKind.h`, so existing users see
   no change.

3. `CompareFlags.h`: move `toString()` and `nullHandlingModeToStr()` to a
   new `CompareFlags.cpp`, which lets the header drop `<fmt/core.h>` and
   the unused `<sstream>`. Also moves `#pragma once` above the includes,
   where it belongs -- it currently sits after them, so the includes were
   re-processed on every inclusion.

Two latent include-what-you-use bugs surfaced once the transitive path
was removed, and are fixed here:

* `Macros.h` uses `std::optional` in `opt_arg_type` / `opt_out_type` but
  never included `<optional>`; it was getting it from `Type.h`.
* `functions/sparksql/MightContain.h` uses `TypePtr` in `initialize()`
  but never included `type/Type.h`.

No behavior change.

Verified: velox_type_test 288, velox_functions_test 1926,
velox_functions_spark_test 707, velox_expression_test 1636 -- 4557 tests,
all passing.
