+++ velox/common/base/CompareFlags.cpp
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) Facebook, Inc. and its affiliates.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "velox/common/base/CompareFlags.h"
+
+#include <fmt/core.h>
+
+namespace facebook::velox {
+
+std::string CompareFlags::nullHandlingModeToStr(NullHandlingMode mode) {
+  switch (mode) {
+    case CompareFlags::NullHandlingMode::kNullAsValue:
+      return "NullAsValue";
+    case CompareFlags::NullHandlingMode::kNullAsIndeterminate:
+      return "NullAsIndeterminate";
+    default:
+      return fmt::format(
+          "Unknown Null Handling mode {}", static_cast<int>(mode));
+  }
+}
+
+std::string CompareFlags::toString() const {
+  return fmt::format(
+      "[NullFirst[{}] Ascending[{}] EqualsOnly[{}] NullHandleMode[{}]]",
+      nullsFirst,
+      ascending,
+      equalsOnly,
+      nullHandlingModeToStr(nullHandlingMode));
+}
+
+} // namespace facebook::velox
