+++ velox/type/TypeKind.h
@@ -342,6 +342,14 @@ constexpr bool is_nested_kind(TypeKind kind) {
       kind == TypeKind::ROW;
 }
 
+/// Evaluates to true only for std::shared_ptr<T>, the physical value type
+/// stored in OPAQUE vectors.
+template <typename>
+struct is_shared_ptr : public std::false_type {};
+
+template <typename T>
+struct is_shared_ptr<std::shared_ptr<T>> : public std::true_type {};
+
 template <TypeKind KIND>
 struct TypeFactory;
